refactor(ooda_loop/orient): split tests to sibling tests_orient.rs (#1266)
- src/ooda_loop/orient.rs: 540 → 124 LOC
- New src/ooda_loop/tests_orient.rs (413 LOC)
- mod registration

Pure refactor: 14 tests pass.
