+++ src/ooda_loop/mod.rs
@@ -16,6 +16,8 @@ mod types;
 
 #[cfg(test)]
 mod tests_observe;
+#[cfg(test)]
+mod tests_orient;
 
 // Re-export all public items so `crate::ooda_loop::X` still works.
 pub use bridge_factory::{bridges_from_state_root, connect_memory};
